server/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.19
psycopg2-binary==2.9.6
pydantic==1.10.12
python-jose==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator
bcrypt==4.0.1
passlib==1.7.4
